// src/enums/PresetsEnum.ts
// 颜色预设枚举
export const enum PresetsEnum {
  // Dust Red
  RED = "red",
  // Volcano
  VOLCANO = "valcano",
  // Sunset Orange
  ORANGE = "orange",
  // Calendula Gold
  GOLD = "gold",
  // Sunrise Yellow
  YELLOW = "yellow",
  // Lime
  LIME = "lime",
  // Polar Green
  GREEN = "green",
  // cyan
  CYAN = "cyan",
  // Daybreak Blue
  DAYBREAK_BLUE = "DaybreakBlue",
  // Geek Blue
  GEEK_BLUE = "GeekBlue",
  // Golden Purple
  GOLDEN_PURPLE = "GoldenPurple",
  // Magenta
  MAGEENTA = "magenta ",
}